Squish into one file, for easier perusing
